Upgrade GDevelop.js jest to v29 and adapt tests
