+++ GDevelop.js/__tests__/GDJSBasicCodeGenerationIntegrationTests.js
@@ -23,12 +23,9 @@ const makeAddOneToObjectTestVariableEvent = (objectName) => ({
 
 describe('libGD.js - GDJS Code Generation integration tests', function () {
   let gd = null;
-  beforeAll((done) =>
-    initializeGDevelopJs().then((module) => {
-      gd = module;
-      done();
-    })
-  );
+  beforeAll(async () => {
+    gd = await initializeGDevelopJs();
+  });
 
   it('does not generate anything for a disabled event and its sub events', function () {
     const serializerElement = gd.Serializer.fromJSObject([
