+++ GDevelop.js/__tests__/Serializer.js
@@ -2,10 +2,9 @@ const initializeGDevelopJs = require('../../Binaries/embuild/GDevelop.js/libGD.j
 
 describe('libGD.js object serialization', function() {
   let gd = null;
-  beforeAll((done) => (initializeGDevelopJs().then(module => {
-    gd = module;
-    done();
-  })));
+  beforeAll(async () => {
+    gd = await initializeGDevelopJs();
+  });
 
   describe('gd.SerializerElement', function() {
     it('should support operations on its value', function() {
