+++ GDevelop.js/package.json
@@ -31,7 +31,7 @@
     "grunt-newer": "^1.2.0",
     "grunt-shell": "^2.1.0",
     "grunt-string-replace": "^1.3.1",
-    "jest": "^23.5.0",
+    "jest": "^29.7.0",
     "shelljs": "^0.8.4",
     "webidl-tools": "github:4ian/webidl-tools#348f9c03afc9d8f278efccdd74543e265a41fd11"
   },
